+++ routes/api/topic.js
@@ -928,7 +928,8 @@ module.exports = function (app) {
                 } else {
                     resObject.sourcePartnerId = null;
                 }
-
+                
+                resObject.favourite = false;
                 resObject.permission = { // TODO: should be plural?
                     level: level
                 };
@@ -1016,6 +1017,10 @@ module.exports = function (app) {
                         WHEN COALESCE(tmup.level, tmgp.level, \'none\') = \'admin\' THEN t."tokenJoin" \
                         ELSE NULL \
                         END as "tokenJoin", \
+                        CASE \
+                        WHEN tf."topicId" = t.id THEN true \
+                        ELSE false \
+                        END as "favourite", \
                         t.categories, \
                         t."endsAt", \
                         t."padUrl", \
@@ -1111,6 +1116,7 @@ module.exports = function (app) {
                         LEFT JOIN "Votes" v \
                                 ON v.id = tv."voteId" \
                     ) AS tv ON (tv."topicId" = t.id) \
+                    LEFT JOIN "TopicFavourites" tf ON tf."topicId" = t.id AND tf."userId" = :userId \
                     ' + join + ' \
                 WHERE t.id = :topicId \
                     AND t."deletedAt" IS NULL \
@@ -1678,6 +1684,10 @@ module.exports = function (app) {
                         WHEN COALESCE(tmup.level, tmgp.level, \'none\') = \'admin\' THEN t."tokenJoin" \
                         ELSE NULL \
                      END as "tokenJoin", \
+                     CASE \
+                        WHEN tf."topicId" = t.id THEN true \
+                        ELSE false \
+                     END as "favourite", \
                      t.categories, \
                      t."sourcePartnerId", \
                      t."sourcePartnerObjectId", \
@@ -1794,6 +1804,7 @@ module.exports = function (app) {
                                 ) AS tcc \
                             GROUP BY tcc."topicId" \
                     ) AS com ON (com."topicId" = t.id) \
+                    LEFT JOIN "TopicFavourites" tf ON tf."topicId" = t.id AND tf."userId" = :userId \
                     ' + join + ' \
                 WHERE ' + where + ' \
                 ORDER BY "order" ASC, t."updatedAt" DESC \