+++ test/api/topic.js
@@ -4577,6 +4577,7 @@ suite('Users', function () {
                     if (err) return done(err);
 
                     delete topic.permission;
+                    delete topic.favourite;
                     topic.padUrl = topic.padUrl.split('?')[0]; // Pad url will not have JWT token as the user gets read-only by default
 
                     var expectedResult = {
@@ -7485,6 +7486,7 @@ suite('Topics', function () {
                         assert.notProperty(topicRead.data, 'events');
 
                         delete topicRead.data.tokenJoin; // Unauth read of Topic should not give out token!
+                        delete topicRead.data.favourite; // Unauth read of Topic should not give out favourite tag value!
 
                         // Also, padUrl will not have authorization token
                         topicRead.data.padUrl = topicRead.data.padUrl.split('?')[0];
@@ -7529,6 +7531,7 @@ suite('Topics', function () {
                                 topicRead.data.permission.level = TopicMemberUser.LEVELS.none;
 
                                 delete topicRead.data.tokenJoin; // Unauth read of Topic should not give out token!
+                                delete topicRead.data.favourite; // Unauth read of Topic should not give out favourite tag value!
 
                                 // Also, padUrl will not have authorization token
                                 topicRead.data.padUrl = topicRead.data.padUrl.split('?')[0];
@@ -7577,6 +7580,7 @@ suite('Topics', function () {
                                 topicRead.data.permission.level = TopicMemberUser.LEVELS.none;
 
                                 delete topicRead.data.tokenJoin; // Unauth read of Topic should not give out token!
+                                delete topicRead.data.favourite; // Unauth read of Topic should not give out favourite tag value!
 
                                 // Also, padUrl will not have authorization token
                                 topicRead.data.padUrl = topicRead.data.padUrl.split('?')[0];
@@ -7631,6 +7635,7 @@ suite('Topics', function () {
                                 topicRead.data.permission.level = TopicMemberUser.LEVELS.none;
 
                                 delete topicRead.data.tokenJoin; // Unauth read of Topic should not give out token!
+                                delete topicRead.data.favourite; // Unauth read of Topic should not give out favourite tag value!
 
                                 // Also, padUrl will not have authorization token
                                 topicRead.data.padUrl = topicRead.data.padUrl.split('?')[0];